--- src/components/util.ts
@@ -1,4 +1,0 @@
-export function timestampToDate(timestamp: string): string {
-  const date = new Date(parseInt(timestamp));
-  return date.toLocaleDateString();
-}
